backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
dependencies = [
    "fastapi==0.124.4",
    "uvicorn==0.38.0",
    "celery==5.6.0",
    "redis==7.1.0",
    "pydantic==2.12.5",
    "httpx==0.28.1",
    "python-dotenv==1.2.1",
    "psycopg2-binary",
    "alembic",
    "sqlalchemy",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
]

[project.optional-dependencies]
dev = [
    "flower==2.0.1",
]
